ETL/requirements.txt
apache-airflow-providers-amazon
pydantic
pandas
pinecone-client
google-generativeai==0.3.1
openai
opencv-python
transformers
pillow
torch